+++ pygenalgo/operators/mutation/polynomial_mutator.py
@@ -85,7 +85,7 @@ class PolynomialMutator(MutationOperator):
             # _end_if_
 
             # Select a random position in the genome.
-            i = self.rng.integers(n_genes)
+            i = self.rng.integers(n_genes, dtype=int)
 
             # Get the old value of the Gene.
             old_value = individual[i].value
